main/arquivos/requirements.txt
flask
requests
dajax
pyserial
configparser
flask-login
gevent
Flask-PyMongo
python-networkmanager
wifi
flask-WTF
passlib